* Refactored response logic * Introduced simplified responses with `IntoResponse` trait
