+++ src/http/response/macros.rs
@@ -1,315 +1,5 @@
-﻿/// Produces an `OK 200` response with plain text or JSON body
-/// 
-/// # Examples
-/// ## plain/text
-/// ```no_run
-/// use volga::ok;
+﻿/// Produces HTTP 404 NOT FOUND response
 ///
-/// ok!("healthy");
-/// ```
-/// ## plain/text without body
-/// ```no_run
-/// use volga::ok;
-///
-/// ok!();
-/// ```
-/// ## JSON
-///```no_run
-/// use volga::ok;
-/// use serde::Serialize;
-///
-/// #[derive(Serialize)]
-/// struct Health {
-///    status: String
-/// }
-///
-/// let health = Health { status: "healthy".into() };
-/// ok!(health);
-/// ```
-/// ## Untyped JSON with custom headers
-///```no_run
-/// use volga::ok;
-/// use serde::Serialize;
-///
-/// #[derive(Serialize)]
-/// struct Health {
-///    status: String
-/// }
-///
-/// ok!({ "health": "healthy" }, [
-///    ("x-api-key", "some api key")
-/// ]);
-/// ```
-#[macro_export]
-macro_rules! ok {
-    // handles ok!()
-    () => {
-        $crate::response!(
-            $crate::http::StatusCode::OK, 
-            $crate::HttpBody::empty(),
-            [
-                ($crate::headers::CONTENT_TYPE, "text/plain")
-            ]
-        )
-    };
-    
-    // handles ok!([("key", "val")])
-    ([ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK, 
-            $crate::HttpBody::empty(),
-            [ $( ($key, $value) ),* ]
-        )
-    };
-    
-    // handles ok!({ json })
-    ({ $($json:tt)* }) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::json($crate::json::json_internal!({ $($json)* })),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-            ]
-        )
-    };
-    
-    // handles ok!({ json }, [("key", "val")])
-    ({ $($json:tt)* }, [ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::json($crate::json::json_internal!({ $($json)* })),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-                $( ($key, $value) ),*
-            ]
-        )
-    };
-    
-    // handles ok!(json)
-    ($var:ident) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::json($var),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-            ]
-        )
-    };
-    
-    // handles ok!(json, [("key", "val")])
-    ($body:expr, [ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::json($body),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-                $( ($key, $value) ),*
-            ]
-        )
-    };
-    
-    // handles ok!("Hello {name}")
-    ($fmt:tt) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::json(format!($fmt)),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-            ]
-        )
-    };
-    
-    // handles ok!(thing.to_string()) or ok!(5 + 5)
-    ($body:expr) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::json($body),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-            ]
-        )
-    };
-    
-    // handles ok!("Hello {}", name)
-    ($($fmt:tt)*) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::json(format!($($fmt)*)),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-            ]
-        )
-    };
-}
-
-/// Produces `OK 200` response with Form Data body
-/// 
-/// # Example
-/// ```no_run
-/// use std::collections::HashMap;
-/// use volga::form;
-///
-/// # async fn dox() -> std::io::Result<()> {
-/// let data = HashMap::from([
-///     ("key", "value")
-/// ]);
-///
-/// form!(data);
-/// # Ok(())
-/// # }
-/// ```
-#[macro_export]
-macro_rules! form {
-    // handles form!({ "key": "value" })
-    ({ $($json:tt)* }) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::form($crate::json::json_internal!({ $($json)* })),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/x-www-form-urlencoded"),
-            ]
-        )
-    };
-    
-    // handles form!({ "key": "value" }, [("key", "val")])
-    ({ $($json:tt)* }, [ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::form($crate::json::json_internal!({ $($json)* })),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/x-www-form-urlencoded"),
-                $( ($key, $value) ),*
-            ]
-        )
-    };
-    
-    // handles form!(object, [("key", "val")])
-    ($body:expr, [ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::form($body),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/x-www-form-urlencoded"),
-                $( ($key, $value) ),*
-            ]
-        )
-    };
-    
-    // handles form!(object)
-    ($body:expr) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::form($body),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/x-www-form-urlencoded"),
-            ]
-        )
-    };
-}
-
-/// Produces `OK 200` response with file body
-/// 
-/// # Examples
-/// ## Default usage
-///```no_run
-/// use volga::file;
-/// use tokio::fs::File;
-///
-/// # async fn dox() -> std::io::Result<()> {
-/// let file_name = "example.txt";
-/// let file_data = File::open(file_name).await?;
-///
-/// file!(file_name, file_data);
-/// # Ok(())
-/// # }
-/// ```
-/// ## Custom headers
-///```no_run
-/// use volga::{file, App};
-/// use tokio::fs::File;
-///
-/// # async fn dox() -> std::io::Result<()> {
-/// let file_name = "example.txt";
-/// let file_data = File::open(file_name).await?;
-/// 
-/// file!(file_name, file_data, [
-///    ("x-api-key", "some api key")
-/// ]);
-/// # Ok(())   
-/// # }
-/// ```
-#[macro_export]
-macro_rules! file {
-    ($file_name:expr, $body:expr) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK, 
-            $crate::HttpBody::wrap_stream($body),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/octet-stream"),
-                ($crate::headers::TRANSFER_ENCODING, "chunked"),
-                ($crate::headers::CONTENT_DISPOSITION, format!("attachment; filename=\"{}\"", $file_name))
-            ]
-        )
-    };
-    
-    ($file_name:expr, $body:expr, [ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK, 
-            $crate::HttpBody::wrap_stream($body),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/octet-stream"),
-                ($crate::headers::TRANSFER_ENCODING, "chunked"),
-                ($crate::headers::CONTENT_DISPOSITION, format!("attachment; filename=\"{}\"", $file_name)),
-                $( ($key, $value) ),*
-            ]
-        )
-    };
-}
-
-/// Produces `OK 200` response with stream body
-/// 
-/// # Examples
-/// ## Default usage
-///```no_run
-/// use volga::{HttpRequest, stream};
-///
-/// # async fn dox(request: HttpRequest) -> std::io::Result<()> {
-/// let boxed_body = request.into_boxed_body();
-/// stream!(boxed_body);
-/// # Ok(())
-/// # }
-/// ```
-/// ## Custom headers
-///```no_run
-/// use volga::{HttpRequest, stream};
-///
-/// # async fn dox(request: HttpRequest) -> std::io::Result<()> {
-/// let boxed_body = request.into_boxed_body();
-/// stream!(boxed_body, [
-///    ("Content-Type", "message/http")
-/// ]);
-/// # Ok(())
-/// # }
-/// ```
-#[macro_export]
-macro_rules! stream {
-    ($body:expr) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK,
-            $crate::HttpBody::new($body)
-        )
-    };
-    ($body:expr, [ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::OK, 
-            $crate::HttpBody::new($body),
-            [ $( ($key, $value) ),* ]
-        )
-    };
-}
-
-/// Produces HTTP 404 NOT FOUND response
-/// 
 /// # Examples
 /// ## Without body
 /// ```no_run
@@ -415,404 +105,16 @@ macro_rules! headers {
     }};
 }
 
-/// Produces a response with specified `StatusCode` with plain text or JSON body
-/// 
-/// # Examples
-/// ## Without body
-/// ```no_run
-/// use volga::status;
-///
-/// status!(404);
-/// ```
-/// ## plain/text body
-/// ```no_run
-/// use volga::status;
-///
-/// status!(401, "Unauthorized!");
-/// ```
-/// ## JSON body
-/// ```no_run
-/// use volga::status;
-/// use serde::Serialize;
-/// 
-/// #[derive(Serialize)]
-/// struct ErrorMessage {
-///     error: String
-/// }
-/// 
-/// let error = ErrorMessage { error: "some error message".into() };
-/// status!(401, error);
-/// ```
-#[macro_export]
-macro_rules! status {
-    ($status:expr, { $($json:tt)* }) => {
-        $crate::response!(
-            $crate::http::StatusCode::from_u16($status).unwrap_or($crate::http::StatusCode::OK),
-            $crate::HttpBody::json($crate::json::json_internal!({ $($json)* })),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-            ]
-        )
-    };
-    
-    ($status:expr) => {
-        $crate::response!(
-            $crate::http::StatusCode::from_u16($status).unwrap_or($crate::http::StatusCode::OK), 
-            $crate::HttpBody::empty(),
-            [
-                ($crate::headers::CONTENT_TYPE, "text/plain")
-            ]
-        )
-    };
-    
-    ($status:expr, [ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::from_u16($status).unwrap_or($crate::http::StatusCode::OK), 
-            $crate::HttpBody::empty(),
-            [ $( ($key, $value) ),* ]
-        )
-    };
-    
-    ($status:expr, $body:expr, [ $( ($key:expr, $value:expr) ),* $(,)? ]) => {
-        $crate::response!(
-            $crate::http::StatusCode::from_u16($status).unwrap_or($crate::http::StatusCode::OK), 
-            $crate::HttpBody::json($body),
-            [ $( ($key, $value) ),* ]
-        )
-    };
-    
-    ($status:expr, $body:expr) => {
-        $crate::response!(
-            $crate::http::StatusCode::from_u16($status).unwrap_or($crate::http::StatusCode::OK),
-            $crate::HttpBody::json($body),
-            [
-                ($crate::headers::CONTENT_TYPE, "application/json"),
-            ]
-        )
-    };
-}
-
 #[cfg(test)]
 mod tests {
     use http_body_util::BodyExt;
-    use std::{path::Path, collections::HashMap};
     use serde::Serialize;
-    use tokio::fs::File;
-    use crate::HttpBody;
-    use crate::test_utils::read_file_bytes;
 
     #[derive(Serialize)]
     struct TestPayload {
         name: String
     }
     
-    #[tokio::test]
-    async fn it_creates_json_ok_response() {
-        let payload = TestPayload { name: "test".into() };
-        let response = ok!(payload);
-        
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.headers().get("Content-Type").unwrap(), "application/json");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_json_from_inline_struct_ok_response() {
-        let response = ok!(TestPayload { name: "test".into() });
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.headers().get("Content-Type").unwrap(), "application/json");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_anonymous_type_json_ok_response() {
-        let response = ok!({ "name": "test" });
-
-        assert!(response.is_ok());
-        
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 200);
-    }
-    
-    #[tokio::test]
-    async fn it_creates_text_ok_response() {
-        let text = "test";
-        let response = ok!(text);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "\"test\"");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_literal_text_ok_response() {
-        let response = ok!("test");
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "\"test\"");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_expr_ok_response() {
-        let response = ok!(5 + 5);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "10");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_number_ok_response() {
-        let number = 100;
-        let response = ok!(number);
-        // this is known issue will be fixed in future releases.
-        //let response = ok!(100);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "100");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_boolean_ok_response() {
-        let response = ok!(true);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "true");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_char_ok_response() {
-        let ch = 'a';
-        let response = ok!(ch);
-        // this is known issue will be fixed in future releases.
-        //let response = ok!('a');
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "\"a\"");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_array_ok_response() {
-        let vec = vec![1,2,3];
-        let response = ok!(vec);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "[1,2,3]");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_formatted_text_ok_response() {
-        let text = "test";
-        let response = ok!("This is text: {}", text);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "\"This is text: test\"");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_interpolated_text_ok_response() {
-        let text = "test";
-        let response = ok!("This is text: {text}");
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "\"This is text: test\"");
-        assert_eq!(response.status(), 200);
-    }
-    
-    #[tokio::test]
-    async fn it_creates_empty_ok_response() {
-        let response = ok!();
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(body.len(), 0);
-        assert_eq!(response.status(), 200);
-    }
-    
-    #[tokio::test]
-    async fn it_creates_file_with_ok_response() {
-        let path = Path::new("tests/resources/test_file.txt");
-        let file_name = path.file_name().and_then(|name| name.to_str()).unwrap();
-        let file = File::open(path).await.unwrap();
-        
-        let response = file!(file_name, file);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = read_file_bytes(&mut response).await;
-        
-        assert_eq!(String::from_utf8_lossy(body.as_slice()), "Hello, this is some file content!");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_file_with_ok_and_custom_headers_response() {
-        let path = Path::new("tests/resources/test_file.txt");
-        let file_name = path.file_name().and_then(|name| name.to_str()).unwrap();
-        
-        let file = File::open(path).await.unwrap();
-
-        let response = file!(file_name, file, [
-            ("x-api-key", "some api key")
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = read_file_bytes(&mut response).await;
-        
-        assert_eq!(String::from_utf8_lossy(body.as_slice()), "Hello, this is some file content!");
-        assert_eq!(response.headers()["x-api-key"], "some api key");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_stream_response() {
-        let path = Path::new("tests/resources/test_file.txt");
-        let file = File::open(path).await.unwrap();
-        let body = HttpBody::wrap_stream(file);
-
-        let response = stream!(body.into_boxed());
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = read_file_bytes(&mut response).await;
-
-        assert_eq!(String::from_utf8_lossy(body.as_slice()), "Hello, this is some file content!");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_stream_response_with_custom_headers() {
-        let path = Path::new("tests/resources/test_file.txt");
-        let file = File::open(path).await.unwrap();
-        let body = HttpBody::wrap_stream(file);
-
-        let response = stream!(body.into_boxed(), [
-            ("x-api-key", "some api key")
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = read_file_bytes(&mut response).await;
-
-        assert_eq!(String::from_utf8_lossy(body.as_slice()), "Hello, this is some file content!");
-        assert_eq!(response.headers()["x-api-key"], "some api key");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_200_response() {
-        let response = status!(200);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(body.len(), 0);
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_200_with_text_response() {
-        let text = "test";
-        let response = status!(200, text);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "\"test\"");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_200_with_json_response() {
-        let payload = TestPayload { name: "test".into() };
-        let response = status!(200, payload);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn it_creates_anonymous_type_200_response_with_json_body() {
-        let response = status!(200, { "name": "test" });
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 200);
-    }
-
     #[tokio::test]
     async fn it_creates_400_response() {
         let response = bad_request!();
@@ -910,112 +212,6 @@ mod tests {
         assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
         assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
     }
-
-    #[tokio::test]
-    async fn it_creates_empty_401_response() {
-        let response = status!(401);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(body.len(), 0);
-        assert_eq!(response.status(), 401);
-    }
-
-    #[tokio::test]
-    async fn it_creates_401_response_with_text_body() {
-        let response = status!(401, "You are not authorized!");
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "\"You are not authorized!\"");
-        assert_eq!(response.status(), 401);
-    }
-
-    #[tokio::test]
-    async fn it_creates_401_response_with_json_body() {
-        let payload = TestPayload { name: "test".into() };
-        let response = status!(401, payload);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 401);
-    }
-
-    #[tokio::test]
-    async fn it_creates_anonymous_type_401_response_with_json_body() {
-        let response = status!(401, { "name": "test" });
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 401);
-    }
-
-    #[tokio::test]
-    async fn it_creates_empty_403_response() {
-        let response = status!(403);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(body.len(), 0);
-        assert_eq!(response.status(), 403);
-    }
-
-    #[tokio::test]
-    async fn it_creates_403_response_with_text_body() {
-        let response = status!(403, "It's forbidden!");
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "\"It's forbidden!\"");
-        assert_eq!(response.status(), 403);
-    }
-
-    #[tokio::test]
-    async fn it_creates_403_response_with_json_body() {
-        let payload = TestPayload { name: "test".into() };
-        let response = status!(403, payload);
-        
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 403);
-    }
-
-    #[tokio::test]
-    async fn it_creates_anonymous_type_403_response_with_json_body() {
-        let response = status!(403, { "name": "test" });
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 403);
-    }
     
     #[tokio::test]
     async fn it_creates_headers() {
@@ -1026,131 +222,6 @@ mod tests {
         
         assert_eq!(headers.get("header 1").unwrap(), "value 1");
         assert_eq!(headers.get("header 2").unwrap(), "value 2")
-    }
-    
-    #[tokio::test]
-    async fn in_creates_text_response_with_custom_headers() {
-        let response = ok!("ok", [
-            ("x-api-key", "some api key"),
-            ("x-req-id", "some req id"),
-        ]);
-        
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-        
-        assert_eq!(String::from_utf8_lossy(body), "\"ok\"");
-        assert_eq!(response.status(), 200);
-        assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
-        assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
-    }
-
-    #[tokio::test]
-    async fn in_creates_text_response_with_empty_custom_headers() {
-        #[allow(unused_mut)]
-        let response = ok!("ok", []);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "\"ok\"");
-        assert_eq!(response.headers().len(), 2);
-        assert_eq!(response.status(), 200);
-    }
-
-    #[tokio::test]
-    async fn in_creates_json_response_with_custom_headers() {
-        let payload = TestPayload { name: "test".into() };
-        let response = ok!(payload, [
-            ("x-api-key", "some api key"),
-            ("x-req-id", "some req id"),
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 200);
-        assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
-        assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
-    }
-
-    #[tokio::test]
-    async fn in_creates_anonymous_json_response_with_custom_headers() {
-        let response = ok!({ "name": "test" }, [
-            ("x-api-key", "some api key"),
-            ("x-req-id", "some req id"),
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 200);
-        assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
-        assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
-    }
-
-    #[tokio::test]
-    async fn it_creates_empty_ok_response_with_headers() {
-        let response = ok!([
-            ("x-api-key", "some api key"),
-            ("x-req-id", "some req id"),
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(body.len(), 0);
-        assert_eq!(response.status(), 200);
-        assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
-        assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
-    }
-
-    #[tokio::test]
-    async fn it_creates_empty_status_response_with_headers() {
-        let response = status!(400, [
-            ("x-api-key", "some api key"),
-            ("x-req-id", "some req id"),
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(body.len(), 0);
-        assert_eq!(response.status(), 400);
-        assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
-        assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
-    }
-
-    #[tokio::test]
-    async fn it_creates_empty_status_response_with_body_and_headers() {
-        let payload = TestPayload { name: "test".into() };
-        let response = status!(406, payload, [
-            ("x-api-key", "some api key"),
-            ("x-req-id", "some req id"),
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(String::from_utf8_lossy(body), "{\"name\":\"test\"}");
-        assert_eq!(response.status(), 406);
-        assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
-        assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
     }
 
     #[tokio::test]
@@ -1184,78 +255,6 @@ mod tests {
         assert_eq!(body.len(), 0);
         assert_eq!(response.status(), 301);
         assert_eq!(response.headers().get("location").unwrap(), url);
-        assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
-        assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
-    }
-
-    #[tokio::test]
-    async fn it_creates_form_data_response() {
-        let data = HashMap::from([
-            ("key", "value"),
-        ]);
-        let response = form!(data);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(response.status(), 200);
-        assert_eq!(String::from_utf8_lossy(body), "key=value");
-        assert_eq!(response.headers().get("Content-Type").unwrap(), "application/x-www-form-urlencoded");
-    }
-
-    #[tokio::test]
-    async fn it_creates_form_data_response_with_headers() {
-        let data = HashMap::from([
-            ("key", "value"),
-        ]);
-        let response = form!(data, [
-            ("x-api-key", "some api key"),
-            ("x-req-id", "some req id"),
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(response.status(), 200);
-        assert_eq!(String::from_utf8_lossy(body), "key=value");
-        assert_eq!(response.headers().get("Content-Type").unwrap(), "application/x-www-form-urlencoded");
-        assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
-        assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
-    }
-
-    #[tokio::test]
-    async fn it_creates_form_data_untyped_response() {
-        let response = form!({ "key": "value" });
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(response.status(), 200);
-        assert_eq!(String::from_utf8_lossy(body), "key=value");
-        assert_eq!(response.headers().get("Content-Type").unwrap(), "application/x-www-form-urlencoded");
-    }
-
-    #[tokio::test]
-    async fn it_creates_form_data_untyped_response_with_headers() {
-        let response = form!({ "key": "value" }, [
-            ("x-api-key", "some api key"),
-            ("x-req-id", "some req id"),
-        ]);
-
-        assert!(response.is_ok());
-
-        let mut response = response.unwrap();
-        let body = &response.body_mut().collect().await.unwrap().to_bytes();
-
-        assert_eq!(response.status(), 200);
-        assert_eq!(String::from_utf8_lossy(body), "key=value");
-        assert_eq!(response.headers().get("Content-Type").unwrap(), "application/x-www-form-urlencoded");
         assert_eq!(response.headers().get("x-api-key").unwrap(), "some api key");
         assert_eq!(response.headers().get("x-req-id").unwrap(), "some req id");
     }
